% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{agregateModulesEnrichment}
\alias{agregateModulesEnrichment}
\title{Modules enrichment taking into account subsets of markers}
\usage{
agregateModulesEnrichment(enrichment.by.module, celltypes, cellsubtypes,
  significanceThreshold)
}
\arguments{
\item{enrichment.by.module}{a matrix with a column for each module in the network and a row for each enrichment type.
Usually, is the matrix generated by \code{\link{getModulesEnrichment}}.}

\item{celltypes}{Vector of charactr with the name of the meta-enrichment types.}

\item{cellsubtypes}{Named list of user-defined lists of genes. Each entry of this list is a vector of a subset of the filenames
used by \code{\link{getModulesEnrichment}}. All the genes included in the files will be considered part of the meta-enrichment whose
name is the name of the list.}

\item{significanceThreshold}{When this value is lower than 1 this function returns the name of the celltype
enrichment of each module which is exclusively enriched by a single celltype with a p-value lower than the
value received by this parameter. In this case, the value of return.processed is ignored.}
}
\value{
a named vector of meta-enrichment names. Each entry is named after the module it has been found. Only those modules including
only one or more significantly p-values from the same meta-enrichment will receive the meta-enrichment name. The rest will have the '-'
character.
}
\description{
This function determine the module enrichment of a GCN taking into account subsets of markers. Each subset is called meta-enrichment.
}
