% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{applyKMeans}
\alias{applyKMeans}
\title{Apply K-means algoritm on a GCN}
\usage{
applyKMeans(net, expr.data, beta = 6, n.iterations = 20,
  net.type = "signed")
}
\arguments{
\item{net}{must be a RDS object storing the network. Networks can be created using \code{\link{createGCN}} function.
We expect a list with objects "moduleColors" and "MEs". The "moduleColors" element is a named vector
whose elements are colors as assigned by WGCNA. The names() function should return the gene names for
each vector's element. The "MEs" is a matrix with a column for each module and a row for each sample.
It contains the module eigengenes. Note that the order of genes in moduleColors is the same as they
appear in columns in expr.data parameter's object.}

\item{expr.data}{can be a full path file name or a data frame with genes in columns and samples in
rows, giving the expression data used to construct the WGCNA network. The function expects genes in the
column order to correspond to the order of nameColors of net.file param.}

\item{beta}{The soft threshold parameter as it was used with WGCNA}

\item{n.iterations}{A max number of iterations to run the k-means}

\item{net.type}{it is used in the same sense as WGCNA}
}
\value{
The network, post-processed with a k-means heuristic including two additional properties
\enumerate{
  \item iterations The number of iterations that the algorithm actually made.
  \item exchanged.genes The number of exchanged genes in the last iteration of the algorithm.
}#'
}
\description{
This function takes a hopefully correctly created WGCN network and, based on gene expression it
modifies the module color assignment by using k-means heuristic.
The description of the parameters show how the function works.
}
