% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCN.R
\name{createGCN}
\alias{createGCN}
\title{Gene Coexpression Network Creation}
\usage{
createGCN(expr.data, net.type = "signed", min.cluster.size = 30,
  n.iterations = 30, save.adj.filename = NULL,
  save.tom.filename = NULL)
}
\arguments{
\item{expr.data}{can be a full path file name or a data frame with genes in columns and samples in
rows, giving the expression data used to construct the WGCNA network.}

\item{net.type}{it is used in the same sense as WGCNA.}

\item{min.cluster.size}{The minimum number of genes used during the dynamic tree cut phase en WGCNA pipeline.}

\item{n.iterations}{A max number of iterations to run the k-means algorithm.}

\item{save.adj.filename}{The name of the file in which to save the adjacency matrix generated by WGCNA.
If you do not want to save this matrix does just leave this parameter with its default value.}

\item{save.tom.filename}{The name of the file in which to save the TOM matrix generated by WGCNA.
If you do not want to save this matrix does just leave this parameter with its default value.}
}
\value{
The network, post-processed with a k-means heuristic. The object returned has the following properties:
\enumerate{
  \item moduleColors a named vector whose elements are colors as assigned by WGCNA. The names() function on this
  propertie should return the gene names for each vector's element. Notice that the order of genes in moduleColors
  is the same as they appear in columns in expr.data parameter's object.
  \item moduleLabels contains a list with the gene names in the same order as in moduleColors.
  \item MEs is a matrix with a column for each module and a row for each sample. It contains the module eigengenes.
  \item beta The soft threshold parameter used to create the adjacency matrix in WGCNA.
  \item type it is used in the same sense as WGCNA.
}#'
}
\description{
This function creates a WGCN network using WGCNA with the additonal step described in https://doi.org/10.1186/s12918-017-0420-6.
}
