% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCN.R
\name{createSCN}
\alias{createSCN}
\title{Secondary Coexpression Network Creation}
\usage{
createSCN(net.filename, expr.data, target.celltype, markers.path = NULL)
}
\arguments{
\item{net.filename}{The full path of a file containing a GCN created with \code{\link{createGCN}}}

\item{expr.data}{can be a full path file name or a data frame with genes in columns and samples in
rows, giving the expression data used to construct the GCN network.}

\item{target.celltype}{A string with the name of the celltype we are interested in. This name should be
one from the list of cell type marker genes defined in the following parameter.}

\item{markers.path}{Folder containing user-defined lists of genes to be used as marker genes to determine
modules enrichment. This is done using WGCNA::userListEnrichment function so they must be in a compatible
format. Gene ids must be expresed using the same format as in the network specified in the first parameter.}
}
\value{
a list with two objects: The GCN network obtained using the expression data filtered to show the
secondary role of genes, and a matrix with that expression data.
}
\description{
This function creates a Secondary Coexpression Network (SCN)
}
