% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{enrichmenEvolution}
\alias{enrichmenEvolution}
\title{Showing changes in enrichment between networks}
\usage{
enrichmenEvolution(primary.net, secondary.net, genes = NULL)
}
\arguments{
\item{primary.net}{A GCN created with \code{\link{createGCN}}. It can be a full path of a file containing it or an R object.}

\item{secondary.net}{A GCN created with \code{\link{createGCN}} from a expression data containing the same genes as the
previous network. It can be a full path of a file containing it or an R object.}

\item{genes}{Subset of genes to be compared. The default value is null and that means all genes will be used in the comparison.}
}
\value{
A dataframe with a row for each gene and the module and enrichment in both networks.
}
\description{
This function compares the enrichment of a set of genes in two networks. Usually, the primary and the secondary network.
}
