% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{enrichmentEvolution}
\alias{enrichmentEvolution}
\title{Resume changes in enrichment for each gene}
\usage{
enrichmentEvolution(primary.net, secondary.net, genes = NULL,
  markers.path = ".")
}
\arguments{
\item{primary.net}{A GCN created with \code{\link{createGCN}}}

\item{secondary.net}{A GCN created with \code{\link{createGCN}}}

\item{genes}{The list of gene's IDs to be used}

\item{markers.path}{Folder containing user-defined lists of genes to be used as marker genes to determine
modules' enrichment. This is done using WGCNA::userListEnrichment function, so they must be in a compatible
format. Gene IDs must be expresed using the same format as in the networks.}
}
\value{
A data frame showing a row for each gene and the module and enrichment of that gene in both networks.
When the module is not enriched a - is shonw. When is enriched by more than a type, they are shown as a list.
}
\description{
This function analyces two networks determining the enrichment of each gene in each network. Both
networks must be made from the same set of genes. The list of genes studied can be provided or all of them
will be analyzed.
}
