% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{Resume Read marker gene files and return a list of marker genes}
\usage{
getMarkerGenes(markers.path = ".")
}
\arguments{
\item{markers.path}{Folder containing user-defined lists of genes to be used as marker genes to determine
modules' enrichment. This is done using WGCNA::userListEnrichment function, so they must be in a compatible
format. Gene IDs must be expresed using the same format as in the expression data specified in the first parameter.}
}
\value{
A list of marker gene lists, one for each file read.
}
\description{
This function read marker gene files and return a list of marker genes.
}
