% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{getModulesEnrichment}
\alias{getModulesEnrichment}
\title{Modules enrichment}
\usage{
getModulesEnrichment(net, markers.path = ".", return.processed = F)
}
\arguments{
\item{net}{A GCN created with \code{\link{createGCN}}. It can be an R object or the full path of a file
containing it. If a filename is provided this function will first look for a file with the same name but
ending in ".enrichment.csv". If the file exists the function will return its content.
This function determine the module enrichment of a GCN using WGCNA::userListEnrichment}

\item{markers.path}{Folder containing user-defined lists of genes to be used as marker genes to determine
modules enrichment. This is done using WGCNA::userListEnrichment function so they must be in a compatible
format. Gene IDs must be expresed using the same format as in the network specified in the first parameter.}

\item{return.processed}{When True the functions returns the -log10 of the p-value obtained replacing the -inf
values by the highest value obtained. When False the function returns the p-values.}
}
\value{
a matrix with a column for each module in the network and a row for each cell type. Values in the matrix
reflects the p-value of the test made to determine if the module is enriched by the cell type or not.
}
\description{
Modules enrichment
}
