% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{removePrimaryEffect}
\alias{removePrimaryEffect}
\title{Cleaning the primary signal}
\usage{
removePrimaryEffect(expr.data, target.enrichment, net = NULL,
  markers.path = NULL)
}
\arguments{
\item{expr.data}{The expression data to be cleaned. It can be a full path file name or a data frame with
genes in columns and samples in rows.}

\item{target.enrichment}{A string with the kind of enrichment that should be removed. This name (or list of
names) should be one from the list of enrichment marker genes' files provided.}

\item{net}{A GCN created with \code{\link{createGCN}} from the same expression data. It can be a full path
of a file containing it or an R object. This network reflects the primary signal and it is used to determine
what modules need to be cleaned in order to find the secondary signal. If no network is provided it will
be created. Therefore, providing one is also a way to reduce the time spent by this function.}

\item{markers.path}{Folder containing user-defined lists of genes to be used as marker genes to determine
modules' enrichment. This is done using WGCNA::userListEnrichment function, so they must be in a compatible
format. Gene IDs must be expresed using the same format as in the expression data specified in the first parameter.}
}
\value{
the expression data filtered that, hopefully, show the secondary role of some genes in the same format
of the input expression data.
}
\description{
This function remove the primary signal of the groups of genes whose primary enrichment is indicated in
the following parameter.
}
