% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCNA.R
\name{varianceEvolution}
\alias{varianceEvolution}
\title{Resume changes in the variance of genes after removing some enrichment signal}
\usage{
varianceEvolution(original.expr.data, cleaned.expr.data, genes = NULL)
}
\arguments{
\item{original.expr.data}{The original expression data. It can be a full path file name or a data frame with
genes in columns and samples in rows.}

\item{cleaned.expr.data}{The expression data obtained by removing some enrichment signal with \code{\link{removePrimaryEffect}}.}

\item{genes}{The list of gene's IDs to be used}
}
\value{
A data frame showing a row for each gene and the variance of that gene in both data sets.
}
\description{
This function compare the variance of expression data before and after removing some enrichment signal using
\code{\link{removePrimaryEffect}}.
}
